refactor: relocate method serialization functions to serial_util.py
