+++ vllm/v1/engine/utils.py
@@ -24,7 +24,6 @@ from vllm.platforms import current_platform
 from vllm.ray.ray_env import get_env_vars_to_copy
 from vllm.utils import (
     get_mp_context,
-    serialize_method_call,
 )
 from vllm.utils.network_utils import (
     get_open_zmq_ipc_path,
@@ -34,6 +33,7 @@ from vllm.utils.network_utils import (
 from vllm.v1.engine.coordinator import DPCoordinator
 from vllm.v1.engine.exceptions import FaultInfo
 from vllm.v1.executor import Executor
+from vllm.v1.serial_utils import serialize_method_call
 from vllm.v1.utils import get_engine_client_zmq_addr, shutdown
 
 if TYPE_CHECKING:
